requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
requests==2.32.3
aiosqlite>=0.19,<0.21
